Refactor Plex API
