+++ resources/lib/plex_functions.py
@@ -1029,40 +1029,3 @@ def GetUserArtworkURL(username):
             url = user.thumb
     LOG.debug("Avatar url for user %s is: %s", username, url)
     return url
-
-
-def transcode_image_path(key, AuthToken, path, width, height):
-    """
-    Transcode Image support
-
-    parameters:
-        key
-        AuthToken
-        path - source path of current XML: path[srcXML]
-        width
-        height
-    result:
-        final path to image file
-    """
-    # external address - can we get a transcoding request for external images?
-    if key.startswith('http://') or key.startswith('https://'):
-        path = key
-    elif key.startswith('/'):  # internal full path.
-        path = 'http://127.0.0.1:32400' + key
-    else:  # internal path, add-on
-        path = 'http://127.0.0.1:32400' + path + '/' + key
-    # This is bogus (note the extra path component) but ATV is stupid when it
-    # comes to caching images, it doesn't use querystrings. Fortunately PMS is
-    # lenient...
-    path = path.encode('utf-8')
-    transcode_path = ('/photo/:/transcode/%sx%s/%s'
-                      % (width, height, utils.quote_plus(path)))
-    transcode_path = transcode_path.decode('utf-8')
-    args = {
-        'width': width,
-        'height': height,
-        'url': path
-    }
-    if AuthToken:
-        args['X-Plex-Token'] = AuthToken
-    return utils.extend_url(transcode_path, args)
